# lib/api-zod/src/generated/types/binanceFuturesBalance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Arbitrage Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { BinanceFuturesBalanceMode } from './binanceFuturesBalanceMode';

export interface BinanceFuturesBalance {
  mode: BinanceFuturesBalanceMode;
  totalUsdt: number;
  availableUsdt: number;
}