# lib/api-zod/src/generated/types/binanceFuturesCloseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Arbitrage Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { BinanceFuturesCloseInputSide } from './binanceFuturesCloseInputSide';

export interface BinanceFuturesCloseInput {
  symbol: string;
  /** Original position side (informational; server closes the actual amount). */
  side?: BinanceFuturesCloseInputSide;
  quantity?: number;
}