# lib/api-zod/src/generated/types/binanceFuturesConfigInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Arbitrage Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { BinanceFuturesConfigInputLiveMode } from './binanceFuturesConfigInputLiveMode';
import type { BinanceFuturesConfigInputMode } from './binanceFuturesConfigInputMode';

export interface BinanceFuturesConfigInput {
  liveMode?: BinanceFuturesConfigInputLiveMode;
  /** Which environment to toggle liveTradingEnabled for. */
  mode?: BinanceFuturesConfigInputMode;
  liveTradingEnabled?: boolean;
}