# lib/api-zod/src/generated/types/binanceFuturesCredentialsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Arbitrage Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { BinanceFuturesCredentialsInputMode } from './binanceFuturesCredentialsInputMode';

export interface BinanceFuturesCredentialsInput {
  mode: BinanceFuturesCredentialsInputMode;
  /** @minLength 1 */
  apiKey: string;
  /** @minLength 1 */
  secret: string;
}