# lib/api-zod/src/generated/types/binanceFuturesEnvStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Arbitrage Scanner API
 * OpenAPI spec version: 0.1.0
 */

export interface BinanceFuturesEnvStatus {
  connected: boolean;
  liveTradingEnabled: boolean;
  /** Masked API key, or null when not connected. */
  key?: string | null;
}