# lib/api-zod/src/generated/types/binanceFuturesOrderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Arbitrage Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { BinanceFuturesOrderInputSide } from './binanceFuturesOrderInputSide';

export interface BinanceFuturesOrderInput {
  /** Asset or pair, e.g. "BTC" or "BTCUSDT". */
  symbol: string;
  side: BinanceFuturesOrderInputSide;
  /** Base-asset quantity (rounded server-side to the symbol step size). */
  quantity: number;
  /** Optional leverage to set before the order. */
  leverage?: number;
}