# lib/api-zod/src/generated/types/binanceFuturesOrderResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Arbitrage Scanner API
 * OpenAPI spec version: 0.1.0
 */

export interface BinanceFuturesOrderResult {
  orderId: string;
  fillPrice: number;
  fillQty: number;
}