# lib/api-zod/src/generated/types/binanceFuturesPositions.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Arbitrage Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { BinanceFuturesPosition } from './binanceFuturesPosition';

export interface BinanceFuturesPositions {
  positions: BinanceFuturesPosition[];
}