# lib/api-zod/src/generated/types/binanceFuturesStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crypto Arbitrage Scanner API
 * OpenAPI spec version: 0.1.0
 */
import type { BinanceFuturesEnvStatus } from './binanceFuturesEnvStatus';
import type { BinanceFuturesStatusLiveMode } from './binanceFuturesStatusLiveMode';

export interface BinanceFuturesStatus {
  liveMode: BinanceFuturesStatusLiveMode;
  testnet: BinanceFuturesEnvStatus;
  mainnet: BinanceFuturesEnvStatus;
}